Address PR review: dim parameter, auto-promotion, and dtype responsibility
RunningNorm: Replace computed reduce_dims with explicit `dim` parameter
(default (0,)) using keepdim=True for flexible reduction semantics.

GaussianPosterior: Replace in-place lerp_/copy_ with non-in-place EMA ops
for automatic dtype promotion from theta. to() preserves current buffer
dtype rather than hardcoding float64.

LossBasedEstimator: Remove .float() casts — each consumer (embedding,
posterior) is responsible for its own dtype handling.

Example 05 embeddings: Add x.float() at entry since nn.Linear expects
float32 input, and replace lerp_ with non-in-place EMA.

+++ falcon/estimators/base.py
@@ -451,7 +451,7 @@ class LossBasedEstimator(StepwiseEstimator):
         embedding = instantiate_embedding(self.embedding_config).to(self.device)
         embedding.eval()
         with torch.no_grad():
-            conditions_device = {k: v.to(self.device).float() for k, v in conditions.items()}
+            conditions_device = {k: v.to(self.device) for k, v in conditions.items()}
             embedded = embedding(conditions_device)
 
         # Create posterior with inferred dimensions (uses latent dim if transformed)
@@ -471,7 +471,7 @@ class LossBasedEstimator(StepwiseEstimator):
         theta = self._to_tensor(batch[f"{self.theta_key}.value"], self.device)
         theta_logprob = self._to_tensor(batch[f"{self.theta_key}.log_prob"])
         conditions = {
-            k: self._to_tensor(batch[f"{k}.value"], self.device).float()
+            k: self._to_tensor(batch[f"{k}.value"], self.device)
             for k in self.condition_keys if f"{k}.value" in batch
         }
 
@@ -614,7 +614,7 @@ class LossBasedEstimator(StepwiseEstimator):
         assert conditions, "Conditions must be provided for sampling."
 
         conditions_device = {
-            k: self._to_tensor(v, self.device).float().expand(num_samples, *v.shape[1:])
+            k: self._to_tensor(v, self.device).expand(num_samples, *v.shape[1:])
             for k, v in conditions.items()
         }
